requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
networkx>=2.6.0
matplotlib>=3.4.0
tqdm>=4.62.0
tensorboard>=2.7.0
cvxpy>=1.2.0
ortools>=9.0.0
gurobipy>=9.5.0
cplex>=22.1.0
coin-or-clp>=1.17.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.0
argparse
json5
pyyaml>=5.4.0
wandb>=0.12.0
gradio>=4.0.0 